requirements.txt
asgiref==3.4.1
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3.3
decorator==5.1.1
Django==3.2.11
django-bootstrap-pagination==1.7.1
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-debug-toolbar-force==0.1.8
django-extensions==3.1.5
django-filter==21.1
django-nine==0.2.5
django-phone-field==1.8.1
django-phonenumber-field==6.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
executing==0.8.2
Faker==11.3.0
flake8==4.0.1
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
ipython==8.0.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.22.3
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.4.1
pluggy==1.0.0
prompt-toolkit==3.0.24
ptyprocess==0.7.0
pure-eval==0.2.1
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.11.2
PyJWT==2.3.0
pyparsing==3.0.8
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
stack-data==0.1.4
text-unidecode==1.3
tomli==1.2.3
traitlets==5.1.1
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
